small fixes in infinf simulations

+++ include/Imigrant.h
@@ -2,15 +2,15 @@
 #define IMIGRANT_H
 
 
-class Imigrant  
+class Imigrant
 {
 	public:
 /* DECLARATION */
-		Imigrant(int input_ch, int size, int input_sp);
-		Imigrant(char origin, int input_ch, int size, int input_sp); /* init Individual as 'A', 'B' or pure "AB" heterozygot*/
-		Imigrant(std::vector<Chromosome>& gamete, int input_sp); /*init Individual by gametes on imput */
+		Imigrant(int input_ch, int size, double input_sp);
+		Imigrant(char origin, int input_ch, int size, double input_sp); /* init Individual as 'A', 'B' or pure "AB" heterozygot*/
+		Imigrant(std::vector<Chromosome>& gamete, double input_sp); /*init Individual by gametes on imput */
 		~Imigrant(){genome.clear();}; /* destructor */
-		
+
 // /* COMPUTIONG METHODS */
 		int getChiasma();
 		void makeGamete(std::vector<Chromosome>& gamete); //this method should be rewriten to accept vector of chromosomes instead of pointer to imigrant
@@ -20,11 +20,11 @@ class Imigrant
 		bool Acheck() const;
 		bool Bcheck() const;
 		int getSelectionPressure() const;
-	
+
 	private:
 		std::vector<Chromosome> genome;
 		int number_of_chromosomes;
-		int selection_pressure;
+		double selection_pressure;
 };
